+++ CHANGELOG.md
@@ -9,15 +9,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- added postgen to run "uv sync"
-- added pregen to validate module name
-- [dev] started adding testing for testing the cookiecutter template
-- add .continue to .gitignore
-- Allow python version to be templated
-- moved around files
+- [dev] Added script for running the tests INSIDE the template
+- Added tests for configuration Settings
+- Added configuration loading to entry point
+- Note: SecretStr is problematic in Settings
+- Added self.save and cls.load for Settings
+- Added configuration.py with Settings class
 - [dev] adjusted ruff to enforce logging vs print
 - Added pydantic dependency and comments for SecretStr
 - Added logging with loguru
+- Allow python version to be templated
+- moved around files
+- [dev] started adding testing for testing the cookiecutter template
+- add .continue to .gitignore
+- added pregen to validate module name
+- added postgen to run "uv sync"
+
+
+
+
+
+
+
+
+
 - [dev] Test for "--version" in CLI
 - This CHANGELOG file to hopefully serve as an evolving example of a
   standardized open source project CHANGELOG.